src/requirements.txt
anyio==3.6.2
autopep8==2.0.1
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
DateTime==5.0
deta==1.1.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
passlib==1.7.4
pycodestyle==2.10.0
pydantic==1.10.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
zope.interface==5.5.2
